attempt to fix abort
